+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 - room prototype function for finding a clear spot in an outward spiral
 - room plans factory now accounts for rooms with no room for a base and instead places the structures in an outward spiral
 
+## Changed
+- updated documentation
+- updated GitHub repo name and url
+
 ## [0.3.2] - 2025-05-07
 
 ## Added